Refactor: extracted all of OasTypeParser to InlineOasTypeParser
